// src/app/modules/myfinance-tsclient-generated/model/instrumentProperties.ts
/**
 * Dac Services
 * Dac Service REST API
 *
 * OpenAPI spec version: 1.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface InstrumentProperties { 
    propertyid: number;
    propertyname: string;
    instrumentid: number;
    value: string;
    valuetype: string;
    validfrom: string;
    validto: string;
}