// src/app/modules/myfinance-tsclient-generated/model/instrumentPropertyListModel.ts
/**
 * Dac Services
 * Dac Service REST API
 *
 * OpenAPI spec version: 1.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { InstrumentProperties } from './instrumentProperties';


export interface InstrumentPropertyListModel { 
    url: string;
    id: string;
    values: Array<InstrumentProperties>;
}